Refactor helper functions into tests/common/utils
